Move download_from_gcs to cobalt/tools
This belongs to Cobalt; tools belongs to Chromium.
