+++ cobalt/tools/download_from_gcs.py
