+++ cobalt/tools/testing/download_from_gcs/clang-format.sha1
