+++ cobalt/tools/testing/download_from_gcs/test_dir/clang-format.sha1
