+++ starboard/shared/starboard/player/BUILD.gn
@@ -34,7 +34,7 @@ static_library("video_dmp") {
 action("player_download_test_data") {
   install_content = true
 
-  script = "//tools/download_from_gcs.py"
+  script = "//cobalt/tools/download_from_gcs.py"
 
   sha_sources = []
   foreach(sha1_file, sha1_files) {
